Update README.md
Learn C++
UFO
It’s game time! And by that we mean, it’s time for you to roll up your sleeves and build a game: UFO.

ufo abduction
The premise:

Invaders from outer space have arrived and are abducting humans using tractor beams. Players must crack the codeword to stop the abduction!

Ok, we’ll admit it’s quite a bit like that classic game, “Hangman”, but with a better premise. Plus, building this command-line game was the Codecademy 2019 Software Engineer Internship Backend Programming Challenge!

Note: This is a more involved project, so please feel free to take breaks as needed.